pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "plugsync"
version = "0.1.0"
description = "Sync skills, agents, and commands from GitHub repos to a local target directory"
requires-python = ">=3.9"
dependencies = ["pyyaml>=6.0"]

[project.scripts]
plugsync = "plugsync.main:main"
